a2a_demo/requirements.txt
google-adk

protobuf~=5.29.4
python-dotenv~=1.1.0